refactor(API): This module now returns a factory function.
In order to leverage the circe-middleware module, this module now returns a factory that returns a
middleware function. This allows for more compartimalized middleware to be created. It also adds the
ability to short-circuit the middleware chain using the values returned by the validate function.
Tests were refactored into single files and documentation was also updated.

This module now returns a factory function.